+++ task/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Prevent trailing slashes in task description which may corrupt
   account at Taskserver
 - Prevent newlines in task descriptions
+- Prevent spaces in new tags
 - Active chips are made bold for easier visual scanning
 - Move task id and status to after the task description